+++ allink_core/apps/locations/abstract_models.py
@@ -5,6 +5,7 @@ from django.utils.translation import ugettext_lazy as _
 from django.db import models
 from django.utils.functional import cached_property
 
+from cms.models.fields import PageField
 from cms.models.fields import PlaceholderField
 from adminsortable.models import SortableMixin
 from parler.models import TranslatableModel, TranslatedField
@@ -273,6 +274,13 @@ class BaseLocationsAppContentPlugin(AllinkBaseAppContentPlugin):
         blank=True,
         help_text=_('Select and arrange specific entries, or, leave blank to select all. (If '
                     'manual entries are selected the category filtering will be ignored.)')
+    )
+    apphook_page = PageField(
+        verbose_name=_(u'Apphook Page'),
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        help_text=_(u'If provided, this Apphook-Page will be used to generate the detail link.'),
     )
 
     zoom_level = models.IntegerField(