+++ allink_core/apps/news/abstract_models.py
@@ -2,8 +2,7 @@
 import datetime
 from django.utils.translation import ugettext_lazy as _
 from django.db import models
-from django.core.cache import cache
-from django.core.cache.utils import make_template_fragment_key
+from cms.models.fields import PageField
 
 from model_utils.models import TimeFramedModel
 from parler.models import TranslatableModel, TranslatedField
@@ -94,6 +93,13 @@ class BaseNewsAppContentPlugin(AllinkBaseAppContentPlugin):
         blank=True,
         help_text=_('Select and arrange specific entries, or, leave blank to select all. (If '
                     'manual entries are selected the category filtering will be ignored.)')
+    )
+    apphook_page = PageField(
+        verbose_name=_(u'Apphook Page'),
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        help_text=_(u'If provided, this Apphook-Page will be used to generate the detail link.'),
     )
 
     def save(self, *args, **kwargs):