+++ allink_core/core/customisation.py
@@ -549,6 +549,7 @@ from django.db import models
 from django.core.cache import cache
 from django.core.cache.utils import make_template_fragment_key
 
+from cms.models.fields import PageField
 from cms.models.pluginmodel import CMSPlugin
 from cms.models.fields import PlaceholderField
 from adminsortable.models import SortableMixin
@@ -627,6 +628,13 @@ class {model_name}AppContentPlugin(AllinkBaseAppContentPlugin):
         blank=True,
         help_text=_('Select and arrange specific entries, or, leave blank to select all. (If '
                     'manual entries are selected the category filtering will be ignored.)')
+    )
+    apphook_page = PageField(
+        verbose_name=_(u'Apphook Page'),
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        help_text=_(u'If provided, this Apphook-Page will be used to generate the detail link.'),
     )
 
     def save(self, *args, **kwargs):